requirements.txt
asgiref==3.8.1
attrs==24.2.0
build==1.0.3
CacheControl==0.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
defusedxml==0.8.0rc2
distlib==0.3.7
Django==5.0.7
django-bootstrap5==24.2
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.2.0
Markdown==3.6
more-itertools==10.1.0
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
pillow==10.4.0
pipenv==2024.0.1
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg==3.2.1
psycopg-binary==3.2.1
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.9.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
rapidfuzz==3.5.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
SecretStorage==3.3.3
shellingham==1.5.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.26.3
zipp==3.17.0
